Change User.version and Organization.version to BigInteger
While running `invoke app.users.sync-edm`:

```
sqlalchemy.exc.DataError: (psycopg2.errors.NumericValueOutOfRange) integer out of range

[SQL: UPDATE "user" SET updated=%(updated)s, version=%(version)s, email=%(email)s, full_name=%(full_name)s, website=%(website)s, affiliation=%(affiliation)s, accepted_user_agreement=%(accepted_user_agreement)s, receive_notification_emails=%(receive_notification_emails)s WHERE "user".guid = %(user_guid)s]
[parameters: {'updated': datetime.datetime(2021, 2, 22, 23, 58, 58, 688511), 'version': 1569621460550, 'email': 'test@example.org', 'full_name': 'Test', 'website': 'https://www.example.org/', 'affiliation': 'Example', 'accepted_user_agreement': True, 'receive_notification_emails': False, 'user_guid': UUID('eb83df6e-8a6e-4ed0-8fa5-23997eb6048d')}]
```

The version we got back is 1569621460550 and it is too big for the postgresql
integer field.

When running `invoke app.organizations.sync-edm`:

```
sqlalchemy.exc.DataError: (psycopg2.errors.NumericValueOutOfRange) integer out of range

[SQL: UPDATE organization SET updated=%(updated)s, version=%(version)s, title=%(title)s, website=%(website)s WHERE organization.guid = %(organization_guid)s]
[parameters: {'updated': datetime.datetime(2021, 2, 23, 0, 23, 38, 64171), 'version': 1587500304510, 'title': 'Example title', 'website': 'https://example.org/', 'organization_guid': UUID('3cde5a0a-a313-4c4c-9057-dfb8c1976fd7')}]
```

The version "1587500304510" is too big for the postgresql integer field.
Changing it to BigInteger fixes this problem.

+++ app/modules/organizations/models.py
@@ -153,7 +153,7 @@ class Organization(db.Model, HoustonModel, OrganizationEDMMixin):
     guid = db.Column(
         db.GUID, default=uuid.uuid4, primary_key=True
     )  # pylint: disable=invalid-name
-    version = db.Column(db.Integer, default=None, nullable=True)
+    version = db.Column(db.BigInteger, default=None, nullable=True)
     title = db.Column(db.String(length=50), nullable=False)
 
     logo_guid = db.Column(db.GUID, default=uuid.uuid4, nullable=True)
